+++ gpu/gpt.h
@@ -51,34 +51,34 @@
 #endif
 
 typedef struct {
-    // Token embedding layer
+    // Token embedding layer (FP16)
     half* d_token_embedding;      // [vocab_size x d_model]
     half* d_token_embedding_grad; // [vocab_size x d_model]
     
-    // Output projection weights
+    // Output projection weights (FP16)
     half* d_W_output;             // [d_model x vocab_size]
     half* d_W_output_grad;        // [d_model x vocab_size]
     
-    // Adam parameters
-    half* d_token_embedding_m;    // First moment for token embeddings
-    half* d_token_embedding_v;    // Second moment for token embeddings
-    half* d_W_output_m;           // First moment for output weights
-    half* d_W_output_v;           // Second moment for output weights
+    // Adam parameters (FP32 for numerical stability)
+    float* d_token_embedding_m;    // First moment for token embeddings
+    float* d_token_embedding_v;    // Second moment for token embeddings
+    float* d_W_output_m;           // First moment for output weights
+    float* d_W_output_v;           // Second moment for output weights
     float beta1;                   // Exponential decay rate for first moment
     float beta2;                   // Exponential decay rate for second moment
     float epsilon;                 // Small constant for numerical stability
     int t;                         // Time step
     float weight_decay;            // Weight decay parameter for AdamW
     
-    // Forward pass buffers
+    // Forward pass buffers (FP16)
     half* d_embedded_input;       // [batch_size x seq_len x d_model]
     half* d_output;               // [batch_size x seq_len x vocab_size]
     
-    // Backward pass buffers
+    // Backward pass buffers (FP16)
     half* d_grad_output;          // [batch_size x seq_len x vocab_size]
 
-    // Loss computation buffer
-    float* d_loss_result;          // [1] - FP32 for accumulation
+    // Loss computation buffer (FP32 for accumulation)
+    float* d_loss_result;          // [1]
     
     // Transformer core
     Transformer* transformer;
